Sort /videos dropdown by stage name
This makes it a LOT easier to find a given stage when adding a new
video, especially since you can now jump to the right area by typing the
stage name. It reamins to be seen if rankings navigation should go the
same way.

We currently re-sort the list on every request, that should probably
change but this is good enough for a PoC.